The janitor is coming back any minute and he's going to throw me out so bear with me.
I introduced a MetaTerm which is a special kind of Term that encapsulates any Term and also keeps track of a Token.
I also introduced a LambdaLexer in addition to the LambdaParser.
The Lexer transforms the String into a Token stream (i.e., List<Token>).
The Parser now operates on this Token stream and gives back a (get ready for awesomeness...) MetaTerm!

The Token now has info about which line, column it came from within the original code, as well as a type (e.g., lambda, right
parenthesis, whitespace, what have you)... That info comes from the Lexer. Now, the nice thing is, that data structure is *complete*.
That means you can go back the original string, including whitespaces, optional parenthesis, and comments.

The Parser builds the Term tree as usual, except it wraps all Terms in those MetaTerms (thus also keeping, but not knowing about, the
line, column etc. meta info). It's really beautiful, but unfortunately, my time is almost up. I can see him coming already.

Now, you might be thinking "but what about the existing visitors, they have to know about the existence of this ugly MetaTerm". The
answer is: No! They don't. In fact, they don't even have to implement visit(MetaTerm) and MetaTerm is *not* part of the abstract lambda
tree. It's a transparent construct that is introduced by the Parser, but it is completely transparent afterwards (i.e., the code that
does the math doesn't even know about MetaTerm). I know it's pretty late, but just trust me alright. It's beautiful.

Finally, when it comes to the highlighting, we've still got the HLE class which now

Let me conclude this epic commit message with a drawing.

[ INPUT ] ---(String)---> [ LEXER ] ---(List<Token>)---> [ PARSER ] ---(MetaTree)---> [ HLE ] ---(String)---> [ INPUT ]

I really gotta run now. Cheers.